test(spec): co-locate temporal tests with parse/column/helpers modules
Split temporal.test.ts into temporal-parse, temporal-column, and
temporal-helpers suites matching the production layout, and update the
test:temporal-parser script paths.
